norm mandatory part

+++ mandatory/includes/libft/Convertion/ft_abs.c
@@ -10,7 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_abs(int value)
+#include <stdint.h>
+
+uint16_t	ft_abs(int value)
 {
 	unsigned int	res;
